+++ packages/components/components/elvis-accordion/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Patch
 
-- Bumped dependency
+- Updated tokens declarations
 
 ## 2.0.6 (02.12.21)